+++ pkg/block/fetcher.go
@@ -691,6 +691,10 @@ func (r *ReplicaLabelRemover) Modify(_ context.Context, metas map[ulid.ULID]*met
 				modified.WithLabelValues(replicaRemovedMeta).Inc()
 			}
 		}
+		if len(l) == 0 {
+			level.Warn(r.logger).Log("msg", "block has no labels left, creating one", r.replicaLabels[0], "deduped")
+			l[r.replicaLabels[0]] = "deduped"
+		}
 		metas[u].Thanos.Labels = l
 	}
 	return nil